Remove nmcp-scripting dependency and define Enum locally
