+++ environment/postInstall
@@ -19,6 +19,5 @@ matplotlib==3.10.7
 # latest allensdk must be installed before nmcp-precomputed to avoid installation error with py3.12
 pip3 install -e git+https://github.com/AllenInstitute/AllenSDK@1bdca3ad884c3a5edea8236161424650603e6f29#egg=allensdk
 pip3 install -e git+https://github.com/AllenNeuralDynamics/nmcp-precomputed@3ffc7a53d26c6900a7c588abae4394a312efc117#egg=nmcp
-pip3 install -e git+https://github.com/AllenNeuralDynamics/nmcp-scripting@585c81574c282800f2781df103471b37cc2399ce#egg=nmcpscripting
 
 echo "All done!"
